Seperated out mutable histogram conecpt, made ConstantHistogram
