+++ project/Build.scala
@@ -106,7 +106,8 @@ object GeotrellisBuild extends Build {
         "org.codehaus.jackson" % "jackson-mapper-asl" % "1.6.1",
         "org.spire-math" %% "spire" % "0.4.0",
         "com.nativelibs4java" %% "scalaxy-loops" % "0.3-SNAPSHOT" % "provided",
-        "net.databinder" %% "dispatch-http" % "0.8.10"),
+        "net.databinder" %% "dispatch-http" % "0.8.10" // for reading args from URLs
+      ),
 
       resolvers ++= Seq(
         "NL4J Repository" at "http://nativelibs4java.sourceforge.net/maven/",
